Cleaning house in the repository
Got rid of a lot of obsolete outputs and junk that is no longer
relevant with the transition to data packages and away from postgres.

Consolidated things we need to keep around mostly under the new
devtools/ directory, which replaces the scripts/ directory, e.g. the
FERC Form 1 to EIA plant/utility ID mapping stuff.

Set aside the self-contained datapkg-msha directory to put in its
own baby repository.

The results/notebooks directory remains, and still needs to be
organized / purged, potentially with some of it being split off
into its own project repositories (as a model for what we would do in
the future for contract analysis work)

Progress toward #330